pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "racing-analysis"
version = "0.1.0"
description = "AI-powered horse racing prediction system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "thomasrocks006-cmyk"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Data Processing
    "polars>=0.19.0",
    "duckdb>=0.9.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pyarrow>=13.0.0",
    
    # Machine Learning
    "catboost>=1.2.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "optuna>=3.3.0",
    "mapie>=0.7.0",  # Conformal prediction
    
    # Feature Engineering
    "scipy>=1.11.0",
    "statsmodels>=0.14.0",
    
    # Web & API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    
    # Data Collection
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    
    # Dashboard
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "altair>=5.1.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
    "jsonschema>=4.19.0",
    
    # Date/Time
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "ipython>=8.16.0",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
]

[project.scripts]
racing = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
